Move child components to separate files
